refactor toLowerCase & toUpperCase fun name and add a trailing new line (#2)
