+++ src/core/pipeline/items/string/transform/mod.rs
@@ -5,5 +5,5 @@ pub mod pad_start;
 pub mod pad_end;
 pub mod ellipsis;
 pub mod to_word_case;
-pub mod uppercase;
-pub mod lowercase;
+pub mod to_lower_case;
+pub mod to_upper_case;
