+++ src/core/pipeline/items/string/transform/to_upper_case.rs
@@ -4,21 +4,21 @@ use crate::core::pipeline::ctx::Ctx;
 use crate::prelude::Value;
 use crate::core::result::Result;
 #[derive(Debug, Clone)]
-pub struct LowerCaseItem {}
+pub struct UpperCaseItem {}
 
-impl LowerCaseItem {
+impl UpperCaseItem {
     pub fn new() -> Self {
         Self {}
     }
 }
 
 #[async_trait]
-impl Item for LowerCaseItem {
+impl Item for UpperCaseItem {
     async fn call<'a>(&self, ctx: Ctx<'a>) -> Result<Ctx<'a>> {
         match ctx.get_value() {
             Value::String(ref s) =>
-                Ok(ctx.with_value(Value::String(s.to_lowercase().to_owned()))),
-            _ => Err(ctx.internal_server_error("lowercase: value is not string"))
+                Ok(ctx.with_value(Value::String(s.to_uppercase().to_owned()))),
+            _ => Err(ctx.internal_server_error("uppercase: value is not string"))
         }
     }
 }
@@ -28,23 +28,23 @@ mod tests {
     use super::*;
 
     #[tokio::test]
-    async fn lowercase_works() {
+    async fn uppercase_works() {
         let ctx = Ctx::initial_state_with_value(Value::String(String::from("AbcD")));
         assert_eq!(
-            LowerCaseItem::new()
+            UpperCaseItem::new()
                 .call(ctx.clone())
                 .await
                 .unwrap()
                 .value
                 .as_str()
                 .unwrap(),
-            "abcd");
+            "ABCD");
     }
 
     #[tokio::test]
     async fn should_check_ctx_value() {
         let ctx = Ctx::initial_state_with_value(Value::Null);
-        let r = LowerCaseItem::new().call(ctx.clone()).await;
+        let r = UpperCaseItem::new().call(ctx.clone()).await;
         assert!(r.is_err());
     }
 }
