+++ src/parser/std/pipeline/global.rs
@@ -20,7 +20,7 @@ use crate::parser::std::pipeline::number::{is_even, is_odd};
 use crate::parser::std::pipeline::object::{assign, ctx_self, is, is_a, object_get, object_previous_value, object_set};
 use crate::parser::std::pipeline::query::query_raw;
 use crate::parser::std::pipeline::string::generation::{cuid, random_digits, slug, uuid};
-use crate::parser::std::pipeline::string::transform::{ellipsis, lowercase, uppercase, pad_end, pad_start, regex_replace, split, trim, to_word_case};
+use crate::parser::std::pipeline::string::transform::{ellipsis, to_lower_case, to_upper_case, pad_end, pad_start, regex_replace, split, trim, to_word_case};
 use crate::parser::std::pipeline::string::validation::{has_prefix, has_suffix, is_alphabetic, is_alphanumeric, is_email, is_hex_color, is_numeric, is_prefix_of, is_secure_password, is_suffix_of, regex_match};
 use crate::parser::std::pipeline::value::{eq, gt, gte, exists, is_false, is_null, is_true, lt, lte, neq, one_of};
 use crate::parser::std::pipeline::vector::{filter, item_at, join, map};
@@ -109,8 +109,8 @@ impl GlobalPipelineInstallers {
         objects.insert("split".to_owned(), split);
         objects.insert("trim".to_owned(), trim);
         objects.insert("toWordCase".to_owned(), to_word_case);
-        objects.insert("lowercase".to_owned(), lowercase);
-        objects.insert("uppercase".to_owned(), uppercase);
+        objects.insert("toLowerCase".to_owned(), to_lower_case);
+        objects.insert("toUpperCase".to_owned(), to_upper_case);
         // string validation
         objects.insert("hasPrefix".to_owned(), has_prefix);
         objects.insert("hasSuffix".to_owned(), has_suffix);
